Adopted from upstream:

- Schema::EntityType is now a proper class (variant<Type,string>), mirroring
  DataKind, so an unrecognised entity_type keeps its literal string.  Call sites
  compare through .type(); the switch in signals.cpp does the same.
- Its helper methods index_column_name()/array_index_name()/metadata_count_key()
  replace our hand-rolled string concatenation in signals.cpp.
- EnumerableProxy is now an abstract base with a pure virtual fetch() and no
  std::function.  Spectra/Chromatograms/WavelengthSpectra drop the std::bind
  ctor.  Ours fetch const (by_id and the range queries are const), so each keeps
  a const fetch_() and adds a thin non-const override.
- Index::find_file(EntityType::Type, DataKind::Type), and the SpectraSource API
  (has_spectra/spectra(source)) that upstream's mzp-inspect now uses.  Mapped
  onto our dual-table reader, which additionally picks profile vs centroid per
  spectrum from the declared representation.

Two upstream defects found and corrected, both worth reporting:

1. entity_type_to_string() returns "wavelength spectrum" WITH A SPACE, and
   index_column_name()/array_index_name()/metadata_count_key() are all built by
   appending to it -- yielding "wavelength spectrum_index", which matches no
   column in any archive.  The underscore spelling is also what the
   specification canonicalised in mzPeak-specification#18.  Kept ours.
2. EntryCmpFn, the new comparator passed to std::ranges::sort, is the
   &&-chained form `a.x < b.x && a.y < b.y && ...`.  That is not a strict weak
   ordering -- two entries differing in opposite directions compare false BOTH
   ways, so they are treated as equivalent and equivalence is not transitive --
   which makes the sort undefined behaviour.  Rewritten lexicographically.  It
   matters beyond tidiness: this sort decides which entry a dimension sees
   first, i.e. which one counts as primary when coalescing.

One deliberate divergence, documented in the code: upstream added
buffer_priority to the dimension grouping key so a primary and a non-primary
entry for one array become separate dimensions and the primary can be preferred.
That is right when the two are REDUNDANT copies.  has_uv.mzpeak's are
COMPLEMENTARY -- detector counts in one column, absorbance in the other, each
null where the other has values -- so splitting them and keeping only the
primary silently drops every row the other carried (it surfaced as "the
intensity array is stored in more than one physical type").  buffer_priority is
excluded from the grouping key so the coalesced path can merge them; that path
still throws if two columns claim the same row.

36/36 in release, debug (-Dwerror) and TSan; run2k digest 9fae6e76aea6a57e.

+++ include/mzpeak/metadata/scans.h
@@ -24,7 +24,7 @@ class Scans final {
 public:
   /// Data for all scans.
   struct Raw {
-    std::vector<float> scan_start_time;
+    std::vector<double> scan_start_time;
   };
 
   /// Default constructor;
